Day 3/Demo 2/requirements.txt
azure-functions
psycopg2-binary